# backend/api/oas-file.yaml
openapi: 3.0.3
info:
  title: AppSciArt API
  version: 1.0.0
  description: API interactions required by AppSciArt platform.
security:
  - bearerAuth: []
...
paths:
...
  /users/{userId}/hackathons:
    x-router-controller: userHackathonsController
    get:
      tags: [User Hackathons]
      summary: Get hackathons in which a user is enrolled
      operationId: getUserEnrolledHackathons
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of hackathons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HackathonList"
  /hackathons/{hackathonId}/participants/me:
    x-router-controller: userHackathonsController
    post:
      tags: [User Hackathons]
      summary: Join the current user to a hackathon
      operationId: joinMeHackathon
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HackathonJoinRequest"
      responses:
        "201":
          description: Enrollment created
        "404":
          description: Resource not found
        default:
          description: Unexpected error
  /hackathons/{hackathonId}/participants/{userId}:
    x-router-controller: userHackathonsController
    post:
      tags: [User Hackathons]
      summary: Join a user to a hackathon
      operationId: joinHackathon
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HackathonJoinRequest"
      responses:
        "201":
          description: Enrollment created
  /users/{userId}/hackathons/{hackathonId}/contributions:
    x-router-controller: userHackathonsController
    get:
      tags: [User Hackathons]
      summary: Get user's contributions in a hackathon
      operationId: getUserHackathonContributions
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User's contributions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contribution"
  /users/{userId}/hackathons/{hackathonId}/clusters/{clusterId}:
    x-router-controller: userHackathonsController
    post:
      tags: [User Hackathons]
      summary: Join a cluster
      operationId: joinCluster
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: clusterId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Cluster enrollment created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    FlowerDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        name:
          type: string
          maxLength: 255
        templateLink:
          type: string
          format: uri
        state:
          type: string
          enum: [IN_BLANK, IN_PROGRESS, IN_REVIEW, PUBLISHED]
        concept:
          type: string
          maxLength: 1500
          nullable: true
        resourcesLink:
          type: string
          format: uri
        seedId:
          type: string
          format: uuid
          minimum: 1
        authors:
          type: array
          items:
            $ref: "#/components/schemas/UserDetails"
        hackathonId:
          type: string
          format: uuid
          minimum: 1
    FruitDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        title:
          type: string
          maxLength: 255
        mainImage:
          type: string
          format: uri
          nullable: true
        state:
          type: string
          enum: [IN_BLANK, IN_PROGRESS, IN_REVIEW, PUBLISHED]
        authorVision:
          type: string
          maxLength: 1500
          nullable: true
        curatorVision:
          type: string
          maxLength: 1500
          nullable: true
        driveLink:
          type: string
          format: uri
          nullable: true
        flowerId:
          type: string
          format: uuid
          nullable: true
    UserDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 50
        surname:
          type: string
          maxLength: 50
        password:
          type: string
          format: password
        gender:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
...
    Hackathon:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        logo:
          type: string
          format: uri
          nullable: true
        type:
          type: string
          enum: [ON_SITE, ONLINE, HYBRID]
        location:
          type: string
          maxLength: 255
          nullable: true
        description:
          type: string
          maxLength: 500
          nullable: true
        editionId:
          type: string
          format: uuid
          minimum: 1
        isVisible:
          type: boolean
        meetLink:
          type: string
          format: uri
          nullable: true
    HackathonList:
      type: array
      items:
        $ref: "#/components/schemas/Hackathon"
    Contribution:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        conceptualMapId:
          type: string
          format: uuid
          minimum: 1
        flower:
          $ref: "#/components/schemas/FlowerDetails"
        fruit:
          $ref: "#/components/schemas/FruitDetails"
...
    HackathonJoinRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items:
            type: string
            enum: ["ARTIST", "TECHNOLOGIST", "SCIENTIST", "OTHER"]
        interests:
          type: string
          nullable: true
...